ci: get every Build CI job compiling and clippy-clean again

+++ kernel-hal/src/bare/arch/x86_64/interrupt.rs
@@ -58,7 +58,7 @@ hal_fn_impl! {
             crate::kstats::note_irq(vector);
             // Cached primary-IRQ reference: every interrupt would otherwise
             // pay an RwLock read + Arc clone via `all_irq().first_unwrap()`.
-            crate::drivers::primary_irq().handle_irq(vector as usize);
+            crate::drivers::primary_irq().handle_irq(vector);
         }
 
         fn msi_alloc_block(requested_irqs: usize) -> HalResult<Range<usize>> {
